+++ libs/application/templates/estate/src/forms/Forms.ts
@@ -7,7 +7,7 @@ import {
   buildSection,
   buildSubmitField,
 } from '@island.is/application/core'
-import { DefaultEvents, Form, FormModes } from '@island.is/application/types'
+import { Form, FormModes } from '@island.is/application/types'
 import { dataCollection } from './Sections/dataCollection'
 import { announcerInfo } from './Sections/announcerInfo'
 import { deceasedSpouse } from './Sections/SpouseOfTheDeceased'
@@ -18,12 +18,8 @@ import { estateDebts } from './Sections/estateDebts'
 import { attachments } from './Sections/attachments'
 import { representative } from './Sections/representative'
 import { approvePrivateDivisionSubmission } from './Sections/approveSubmission'
-import { overview } from './Overview/overview'
-import { overviewPrivateDivision } from './Overview/privateOverviewSection'
+import { overview } from './Overview'
 import { estateWithoutAssets } from './Sections/estateWithoutAssets'
-import { m } from '../lib/messages'
-import { deceasedInfoFields } from './Sections/deceasedInfoFields'
-import { YES } from '../lib/constants'
 
 /* EINKASKIPTI */
 
@@ -44,7 +40,7 @@ export const privateDivisionForm: Form = buildForm({
     attachments,
     representative,
     approvePrivateDivisionSubmission,
-    overviewPrivateDivision,
+    overview,
   ],
 })
 
@@ -95,60 +91,5 @@ export const officialDivisionForm: Form = buildForm({
   mode: FormModes.DRAFT,
   renderLastScreenButton: true,
   renderLastScreenBackButton: true,
-  children: [
-    dataCollection,
-    announcerInfo,
-    buildSection({
-      id: 'overviewDivisionOfEstate',
-      title: m.overviewTitle,
-      children: [
-        buildMultiField({
-          id: 'overviewDivisionOfEstate',
-          title: m.overviewTitle,
-          description: m.overviewSubtitleDivisionOfEstate,
-          children: [
-            buildDividerField({}),
-            buildDescriptionField({
-              id: 'deceasedHeader',
-              title: m.theDeceased,
-              titleVariant: 'h3',
-              marginBottom: 2,
-              space: 'gutter',
-            }),
-            ...deceasedInfoFields,
-            buildDescriptionField({
-              id: 'space0',
-              title: '',
-              space: 'containerGutter',
-            }),
-            buildCheckboxField({
-              id: 'confirmAction',
-              title: '',
-              large: true,
-              backgroundColor: 'blue',
-              defaultValue: [],
-              options: [
-                {
-                  value: YES,
-                  label: m.divisionOfEstateConfirmActionCheckbox.defaultMessage,
-                },
-              ],
-            }),
-            buildSubmitField({
-              id: 'officialDivisionForm.submit',
-              title: '',
-              refetchApplicationAfterSubmit: true,
-              actions: [
-                {
-                  event: DefaultEvents.SUBMIT,
-                  name: m.submitApplication,
-                  type: 'primary',
-                },
-              ],
-            }),
-          ],
-        }),
-      ],
-    }),
-  ],
+  children: [dataCollection, announcerInfo, overview],
 })
