+++ libs/application/templates/estate/src/forms/Overview.ts
@@ -1,0 +1,166 @@
+import {
+  buildCheckboxField,
+  buildDescriptionField,
+  buildDividerField,
+  buildMultiField,
+  buildSection,
+  buildSubmitField,
+  getValueViaPath,
+} from '@island.is/application/core'
+import { DefaultEvents } from '@island.is/application/types'
+import { m } from '../lib/messages'
+import { commonOverviewFields } from './OverviewSections/commonFields'
+import { overviewAssetsAndDebts } from './OverviewSections/overviewAssetsAndDebts'
+import { overviewAttachments } from './OverviewSections/overviewAttachments'
+import { overviewConfirmAction } from './OverviewSections/overviewConfirmAction'
+import { EstateTypes, YES, NO } from '../lib/constants'
+import { deceasedInfoFields } from './Sections/deceasedInfoFields'
+import { representativeOverview } from './OverviewSections/representative'
+
+export const overview = buildSection({
+  id: 'overviewEstateDivision',
+  title: m.overviewTitle,
+  children: [
+    /* Einkaskipti */
+    buildMultiField({
+      id: 'overviewPrivateDivision',
+      title: m.overviewTitle,
+      description: m.overviewSubtitleDivisionOfEstateByHeirs,
+      condition: (answers) =>
+        getValueViaPath(answers, 'selectedEstate') ===
+        EstateTypes.divisionOfEstateByHeirs,
+      children: [
+        ...commonOverviewFields,
+        ...overviewAssetsAndDebts,
+        ...overviewAttachments,
+        ...representativeOverview,
+      ],
+    }),
+
+    /* Seta í óskiptu búi og Eignalaust bú með eignum */
+    buildMultiField({
+      id: 'overviewEstateDivision',
+      title: m.overviewTitle,
+      description: (application) => {
+        const selectedEstate = getValueViaPath(
+          application.answers,
+          'selectedEstate',
+        )
+        return selectedEstate === EstateTypes.permitForUndividedEstate
+          ? m.overviewSubtitlePermitToPostpone
+          : selectedEstate === EstateTypes.estateWithoutAssets
+          ? m.overviewSubtitleWithoutAssets
+          : m.overviewSubtitleDivisionOfEstate
+      },
+      condition: (answers) =>
+        getValueViaPath(answers, 'selectedEstate') ===
+        EstateTypes.estateWithoutAssets
+          ? getValueViaPath(
+              answers,
+              'estateWithoutAssets.estateAssetsExist',
+            ) === YES
+          : getValueViaPath(answers, 'selectedEstate') ===
+            EstateTypes.permitForUndividedEstate
+          ? true
+          : false,
+      children: [
+        ...commonOverviewFields,
+        ...overviewAssetsAndDebts,
+        ...overviewAttachments,
+        ...overviewConfirmAction,
+        buildSubmitField({
+          id: 'estateDivisionSubmit.submit',
+          title: '',
+          refetchApplicationAfterSubmit: true,
+          actions: [
+            {
+              event: DefaultEvents.SUBMIT,
+              name: m.submitApplication,
+              type: 'primary',
+            },
+          ],
+        }),
+      ],
+    }),
+
+    /* Eignalaust bú án eigna */
+    buildMultiField({
+      id: 'overviewWithoutAssets',
+      title: m.overviewTitle,
+      description: m.overviewSubtitleWithoutAssets,
+      condition: (answers) =>
+        getValueViaPath(answers, 'selectedEstate') ===
+          EstateTypes.estateWithoutAssets &&
+        getValueViaPath(answers, 'estateWithoutAssets.estateAssetsExist') ===
+          NO,
+      children: [
+        ...commonOverviewFields,
+        ...overviewAttachments,
+        ...overviewConfirmAction,
+        buildSubmitField({
+          id: 'estateDivisionSubmit.submit',
+          title: '',
+          refetchApplicationAfterSubmit: true,
+          actions: [
+            {
+              event: DefaultEvents.SUBMIT,
+              name: m.submitApplication,
+              type: 'primary',
+            },
+          ],
+        }),
+      ],
+    }),
+
+    /* Opinber Skipti */
+    buildMultiField({
+      id: 'overviewDivisionOfEstate',
+      title: m.overviewTitle,
+      description: m.overviewSubtitleDivisionOfEstate,
+      condition: (answers) =>
+        getValueViaPath(answers, 'selectedEstate') ===
+        EstateTypes.officialDivision,
+      children: [
+        buildDividerField({}),
+        buildDescriptionField({
+          id: 'deceasedHeader',
+          title: m.theDeceased,
+          titleVariant: 'h3',
+          marginBottom: 2,
+          space: 'gutter',
+        }),
+        ...deceasedInfoFields,
+        buildDescriptionField({
+          id: 'space0',
+          title: '',
+          space: 'containerGutter',
+        }),
+        buildCheckboxField({
+          id: 'confirmAction',
+          title: '',
+          large: true,
+          backgroundColor: 'blue',
+          defaultValue: [],
+          options: [
+            {
+              value: YES,
+              label: m.divisionOfEstateConfirmActionCheckbox.defaultMessage,
+            },
+          ],
+        }),
+        buildSubmitField({
+          id: 'officialDivisionForm.submit',
+          title: '',
+          refetchApplicationAfterSubmit: true,
+          actions: [
+            {
+              event: DefaultEvents.SUBMIT,
+              name: m.submitApplication,
+              type: 'primary',
+            },
+          ],
+        }),
+      ],
+    }),
+  ],
+})
