+++ libs/application/templates/estate/src/forms/OverviewSections/commonFields.ts
