+++ libs/application/templates/estate/src/forms/OverviewSections/overviewAttachments.ts
@@ -61,7 +61,7 @@ export const overviewAttachments = [
     },
   }),
   buildDescriptionField({
-    id: 'space4',
+    id: 'spaceAttachments',
     title: '',
     space: 'gutter',
   }),
