+++ libs/application/templates/estate/src/forms/OverviewSections/overviewWithoutAssets.ts
